Refactor(settings): Allow sync interval change without internet
This commit modifies the Settings screen to allow users to change the sync interval even when there is no internet connection. Previously, the sync interval setting was disabled if there was no internet or if a sync was in progress. Now, it is only disabled if a sync is currently in progress.

Additionally, added `milestone*.png` to `.gitignore`.